requirements.txt
Django >=1.5
dj-database-url >=0.2.1
psycopg2 >=2.4.6
gunicorn >=0.17.2
django_countries >=1.5
     django-heroku-postgresify
     django-smart-selects
     South